Moving pagination to helpers to make it reusable, implementing helpers in GET transactions
